# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Patient Tracker App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/hospitals": {
            "get": {
                "tags": [
                    "hospitals"
                ],
                "summary": "get all hospitals paginated",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "coe , referal",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "hospitals"
                ],
                "summary": "add new hospital",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "coe , referal",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "rheuma",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "crdio",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "pulmo",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "pah_expert",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "rhc",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "rwe",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "echo",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "pah_attentive",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "city_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "country_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/hospitals/{hospital}": {
            "get": {
                "tags": [
                    "hospitals"
                ],
                "summary": "get single hospitals",
                "parameters": [
                    {
                        "name": "hospital",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "hospitals"
                ],
                "summary": "update hospital",
                "parameters": [
                    {
                        "name": "hospital",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "coe , referal",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "rheuma",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "crdio",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "pulmo",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "pah_expert",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "rhc",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "rwe",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "echo",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "pah_attentive",
                        "in": "formData",
                        "description": "0 , 1",
                        "type": "integer"
                    },
                    {
                        "name": "city_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "country_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "_method:put",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "hospitals"
                ],
                "summary": "Delete Hospital",
                "parameters": [
                    {
                        "name": "hospital",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}